% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{addMapboxGL}
\alias{addMapboxGL}
\title{Add a Mapbox GL layer to a Leaflet map}
\usage{
addMapboxGL(map, options)
}
\description{
Add a Mapbox GL layer to a Leaflet map
}
\examples{
leaflet(quakes) \%>\%
  addMapboxGL(options = mapboxOptions(
    style = "mapbox://styles/mapbox/light-v9"
  )) \%>\%
  addMarkers()
}
